RendezvOS: harden port table backend + cache semantics
Ensure port table lifetime/sentinel correctness and robust ht rehash/shrink under IPC-heavy workloads.

Invariants:
- Table lock protects mutations to slots/ht/free_head/live_ports/ht_tombs.
- Freelist empty sentinel is type-consistent with u64 (free_head).
- Slot reuse bumps slot->gen before the slot is made available.
- ht rehash/shrink uses two-phase commit: build fully, then swap.
- Lookup/resolve returns exactly one live Message_Port ref on success.
- Port table teardown keeps allocator valid until the final table free.
- Thread port cache eviction age is wrap-safe (u16 modular comparison).

Failure paths:
- Grow/rehash/shrink allocation failures free only the newly built structure; old ht remains valid.
- Register abort on ref_get failure rolls back mapping and slot ownership.
- Cache token resolve failure evicts only the stale cache entry.

Lock/refcount order:
- All backend slot/ht/free_head mutations occur under table lock.
- Refs are acquired/put according to backend contract (unlock after stable state; put exactly once).

Teardown behavior:
- fini drains/unregisters remaining registered ports and then frees slots/ht buffers via the table allocator.

Verification:
- Lints checked for modified files.
- Port robustness and IPC tests (including ht shrink scenario) pass as reported.

Made-with: Cursor

+++ modules/test/single_ipc_test.c
@@ -194,8 +194,8 @@ static void* ipc_multi_round_sender_thread(void* arg)
                                           sizeof(multi_ipc_send_payload));
                 if (!payload) {
                         pr_error(
-                                "[single_ipc_multi_round_test] sender: payload alloc failed at round %llu\n",
-                                (unsigned long long)i);
+                                "[single_ipc_multi_round_test] sender: payload alloc failed at round %u\n",
+                                (u64)i);
                         break;
                 }
 
@@ -227,8 +227,8 @@ static void* ipc_multi_round_sender_thread(void* arg)
                         (i64)i, 16, &payload_ptr, free_msgdata_ref_default);
                 if (!msgdata) {
                         pr_error(
-                                "[single_ipc_multi_round_test] sender: create_message_data failed at round %llu\n",
-                                (unsigned long long)i);
+                                "[single_ipc_multi_round_test] sender: create_message_data failed at round %u\n",
+                                (u64)i);
                         percpu(kallocator)->m_free(percpu(kallocator), payload);
                         break;
                 }
@@ -236,8 +236,8 @@ static void* ipc_multi_round_sender_thread(void* arg)
                 Message_t* msg = create_message_with_msg(msgdata);
                 if (!msg) {
                         pr_error(
-                                "[single_ipc_multi_round_test] sender: create_message_with_msg failed at round %llu\n",
-                                (unsigned long long)i);
+                                "[single_ipc_multi_round_test] sender: create_message_with_msg failed at round %u\n",
+                                (u64)i);
                         ref_put(&msgdata->refcount, free_msgdata_ref_default);
                         break;
                 }
@@ -249,8 +249,8 @@ static void* ipc_multi_round_sender_thread(void* arg)
 
                 if (enqueue_msg_for_send(msg) != REND_SUCCESS) {
                         pr_error(
-                                "[single_ipc_multi_round_test] sender: enqueue_msg_for_send failed at round %llu\n",
-                                (unsigned long long)i);
+                                "[single_ipc_multi_round_test] sender: enqueue_msg_for_send failed at round %u\n",
+                                (u64)i);
                         ref_put(&msg->ms_queue_node.refcount, free_message_ref);
                         break;
                 }
@@ -258,8 +258,8 @@ static void* ipc_multi_round_sender_thread(void* arg)
 
                 if (send_msg(port) != REND_SUCCESS) {
                         pr_error(
-                                "[single_ipc_multi_round_test] sender: send_msg failed at round %llu\n",
-                                (unsigned long long)i);
+                                "[single_ipc_multi_round_test] sender: send_msg failed at round %u\n",
+                                (u64)i);
                         break;
                 }
